Port sensors, loader, unloader and exec packages to windows
In order to port sensor package on Windows, adding exec parser and loader / unloader for Windows
Tried to keep common functions between two OS in pkg/sensors/load.go.
The windows implementation of some functions in load_windows.go looks like Linux implementation but is different because of unavailability of some features on windows like bpffs and btf.
Also, the sensors.load() function in Windows is a work in progress and we might deprioritize code-reuse in favour of an ability to make future changes independently.
